Deploy Singularity v3.8.3  * update singularity* packages from upstream  * enable deployment of singularity@3.8.3  * note that permissions/setuid part will be handled    manually for now (as part of PoC phase)
